requirements.txt
click>=8.0.0
openai>=1.0.0
anthropic>=0.18.0
zep-cloud>=0.3.0
zep-python>=0.40.0
google-generativeai>=0.3.0
groq>=0.4.0
mistralai>=1.0.0
together>=1.0.0
cohere>=4.0.0
httpx>=0.24.0

